+++ src/base/widget_animator_manager.h
@@ -39,7 +39,7 @@ struct _widget_animator_manager_t {
 
   /*private*/
   uint32_t timer_id;
-  uint32_t last_dispatch_time;
+  uint64_t last_dispatch_time;
   widget_animator_t* first;
 };
